Splitted the code to follow strictly only 1 feature per member
