+++ public/index.html
@@ -10,7 +10,7 @@
   <body>
     <div class="app-container">
       <header>
-        <h1>Library System</h1>
+        <h1>LibSys</h1>
         <div id="user-section">
           <span id="current-user"></span>
           <button id="logout-btn" class="hidden">Logout</button>
@@ -92,10 +92,14 @@
       </div>
     </div>
 
-    <script src="auth.js"></script>
+    <script src="js/auth.js"></script>
     <script src="js/login.js"></script>
     <script src="js/bryan.js"></script>
     <script src="js/jonathan.js"></script>
     <script src="js/william.js"></script>
+    <script src="js/loadBooks.js"></script>
+    <script src="js/viewBook.js"></script>
+    <script src="js/editBooks.js"></script>
+    <script src="js/adminButtons.js"></script>
   </body>
 </html>
