+++ utils/editBookUtil.js
@@ -1,0 +1,59 @@
+const fs = require('fs').promises;
+const path = require('path');
+
+const BOOK_FILE = path.join(__dirname, 'books.json');
+const TEMPLATE_FILE = path.join(__dirname, 'books.template.json');
+
+// read books from JSON file
+async function readBooksFile() {
+  try {
+    const raw = await fs.readFile(BOOK_FILE, 'utf8');
+    return JSON.parse(raw);
+  } catch (e) {
+    // fallback template
+    try {
+      const tpl = await fs.readFile(TEMPLATE_FILE, 'utf8');
+      await fs.writeFile(BOOK_FILE, tpl, 'utf8');
+      return JSON.parse(tpl);
+    } catch {
+      await fs.writeFile(BOOK_FILE, JSON.stringify({ books: [] }, null, 2), 'utf8');
+      return { books: [] };
+    }
+  }
+}
+
+// write books to JSON file
+async function writeBooksFile(data) {
+  await fs.writeFile(BOOK_FILE, JSON.stringify(data, null, 2), 'utf8');
+}
+
+// PUT /books/:title
+// body: { title, author, content }
+async function updateBook(req, res) {
+  try {
+    const originalTitle = req.params.title;
+    const { title, author, content } = req.body;
+
+    if (!originalTitle || !title || !author || !content) {
+      return res.status(400).json({ error: 'original title, title, author, content required' });
+    }
+
+    const data = await readBooksFile();
+    data.books = data.books || [];
+
+    const idx = data.books.findIndex(b => b.title === originalTitle);
+    if (idx === -1) return res.status(404).json({ error: 'Book not found' });
+
+    data.books[idx] = { ...data.books[idx], title, author, content };
+    await writeBooksFile(data);
+
+    res.json({ message: 'Book updated', book: data.books[idx] });
+  } catch (e) {
+    console.error('updateBook error', e);
+    res.status(500).json({ error: 'Failed to update book' });
+  }
+}
+
+module.exports = {
+  updateBook
+};
